Work around a limitiation in nmcli in setting the mtu
nmcli doesn't appear to want to set the link mtu for a link
that has a disabled ip configuration.  That is, it doesn't
want to the equivalent of an `ip link set <dev> mtu ...`
Adding an explict work around via the shell.

Note this will will reset to the less efficient 1500 default
mtu after reboot so unfortunately isn't permanent, even
with the mtu recorded in the on disk config.